requirements.txt
beautifulsoup4==4.8.1
Flask-Mail==0.9.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
urllib3==1.25.6
vcrpy==2.1.0